+++ template/ts/files/package.json
@@ -16,7 +16,7 @@
   "homepage": "https://github.com/helix-bridge/helixconf",
   "scripts": {
     "build": "rm -rf dist && tsc",
-    "test": "jest"
+    "test": "jest --verbose --expand --ci --coverage"
   },
   "author": "HelixBridge",
   "license": "MIT",
@@ -24,7 +24,7 @@
     "@types/jest": "^29.5.12",
     "ethers": "^6.13.2",
     "jest": "^29.7.0",
-    "ts-jest": "^29.2.3",
+    "ts-jest": "^29.2.4",
     "ts-node": "^10.9.2",
     "typescript": "^5.5.4"
   }
