+++ template/ts/files/tests/_base.ts
@@ -1,0 +1,3 @@
+import {Onlinechain} from "./common/onlinechain";
+
+export const oc = new Onlinechain();
