+++ template/ts/files/tests/common/testsource.ts
@@ -1,4 +1,4 @@
-import {HelixChain, _NetworkType, HelixChainConf, ChainToken} from "../../src";
+import {HelixChain, _NetworkType, HelixChainConf, ChainToken, ChainCouple, ChainMessager} from "../../src";
 
 
 export enum Category {
@@ -19,6 +19,18 @@ export interface IsSkipOptions {
   token?: ChainToken,
 }
 
+export interface TestChainToken extends ChainToken {
+  _chain: string
+}
+
+export interface TestChainCouple extends ChainCouple {
+  _chain: string
+}
+
+export interface TestChainMessager extends ChainMessager {
+  _chain: string
+}
+
 export class TestSource {
   public static chains(options?: TestSourceChainsOptions): HelixChainConf[] {
     const chains = HelixChain.chains({network: options?.network ?? 'mainnets'});
@@ -31,6 +43,43 @@ export class TestSource {
         return chains;
     }
   }
+
+  public static tokens(options?: TestSourceChainsOptions): TestChainToken[] {
+    const chains = TestSource.chains(options);
+    const tokens = [];
+    for (const chain of chains) {
+      const chainTokens = chain.tokens.map(item => {
+        return {...item, _chain: chain.code,} as TestChainToken
+      });
+      tokens.push(...chainTokens);
+    }
+    return tokens;
+  }
+
+  public static couples(options?: TestSourceChainsOptions): TestChainCouple[] {
+    const chains = TestSource.chains(options);
+    const couples = [];
+    for (const chain of chains) {
+      const chainCouples = chain.couples.map(item => {
+        return {...item, _chain: chain.code} as TestChainCouple
+      });
+      couples.push(...chainCouples);
+    }
+    return couples;
+  }
+
+  public static messagers(options?: TestSourceChainsOptions): TestChainMessager[] {
+    const chains = TestSource.chains(options);
+    const messagers = [];
+    for (const chain of chains) {
+      const chainMessagers = chain.messagers.map(item => {
+        return {...item, _chain: chain.code} as TestChainMessager
+      });
+      messagers.push(...chainMessagers);
+    }
+    return messagers;
+  }
+
 
   public static isSkip(options: IsSkipOptions) {
     const {category, chain} = options;
